fix(database): 修复端点表字段类型不匹配问题
- 将Endpoint结构体中的ID和PlatformID字段从int64改为uint类型
- 更新migrate_endpoint.go中平台ID赋值逻辑，移除多余类型转换
- 修改query/endpoints.gen.go中生成的字段类型定义，确保与数据库模型一致

+++ database/migrate_endpoint.go
@@ -73,7 +73,7 @@ func migrateEndpoints(db *gorm.DB) error {
 	for _, platform := range platforms {
 		// 创建默认端点
 		endpoint := types.Endpoint{
-			PlatformID:      int64(platform.ID),
+			PlatformID:      platform.ID,
 			EndpointType:    platform.Provider,
 			EndpointVariant: platform.Variant,
 			Path:            "",
